113 create robust scoring system for gateway tunnels (#114)
- Structure to build and assess substrate node metrics, for gateway load balancing
- HTTP Serviceable scoring Based on:
  - Cpu usage
  - Serviceable assets cached in DAG
  - Serviceable runtime shadows created 
  - Serviceable memory allocation required to call vs available memory
  - Serviceable call history
 - Optimizations with http serviceable creation 
  -  Separate serviceable creation and provisioning serviceable execution requirements to minimize unnecessary operations

+++ vm/vars.go
@@ -4,7 +4,7 @@ import "time"
 
 var (
 	InstanceMaxRequests   int           = 1024 * 64
-	InstanceMaxError      int           = 10
+	InstanceMaxError      int64         = 10
 	InstanceErrorCoolDown time.Duration = 30 * time.Minute
 
 	ShadowBuff                        = 10
